no 32

+++ lib/datasets/Columbia.py
@@ -135,11 +135,11 @@ class Columbia_pascal_voc(imdb):
 
         num_objs = len(objs)
 
-        boxes = np.zeros((num_objs, 4), dtype=np.uint32)
-        gt_classes = np.zeros((num_objs), dtype=np.int32)
-        overlaps = np.zeros((num_objs, self.num_classes), dtype=np.float32)
+        boxes = np.zeros((num_objs, 4), dtype=np.uint64)
+        gt_classes = np.zeros((num_objs), dtype=np.int64)
+        overlaps = np.zeros((num_objs, self.num_classes), dtype=np.float64)
         # "Seg" area for pascal is just the box area
-        seg_areas = np.zeros((num_objs), dtype=np.float32)
+        seg_areas = np.zeros((num_objs), dtype=np.float64)
 
         # Load object bounding boxes into a data frame.
         for ix, obj in enumerate(objs):
